Add Makefile to the project
Makefiles are always great to make it easier
to compile the project and this is enough reason
for them to exist.

assets/ directory was included to store the giff
demo on how to execute the project. It is shown
in README.md file.

README.md file now explains how to compile and
execute the project.